requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
duckdb>=0.9.0

# AI and ML
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-experimental>=0.0.50

# Web app and visualization
streamlit>=1.28.0
plotly>=5.15.0
seaborn>=0.12.0
matplotlib>=3.7.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.7.0

# Development and testing
jupyter>=1.0.0
ipywidgets>=8.0.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Optional: For presentation
jupyterlab>=4.0.0
voila>=0.5.0